Monitor texture setup and can show space scene
If I manually replace the screen texture ID it will display the probe scene. It needs work because the aspect ratio relative to what I UVed in Blender is off, and I forgot that since I only have a single camera, the camera controls are manipulating both scenes (space and workstation).