Switched to a Game Module Interface model.
I ported Epoch over to be a Game Module, the other test project is not fully ported. This gets around the architectural issue of needing multiple main.cpps and main functions in the SLN. Instead, multiple game modules can exist within the same project / solution and you can just register one (or more?) with the engine. The engine then runs the module via the interface.