Object component model foundational classes
Object and Object Component are base classes defining the owner/container object with component attachment and retrieval.

Entity and Scene Component define an in-world object with a world transform.

Entity will be a "renderable thing in a scene". The existing Camera will probably be converted to derive from NObject, and contain a NSceneComponent.